test(android): a Byte is not an Int to JUnit's assertEquals

+++ android/app/src/test/kotlin/com/webyar/operator/ui/components/AttachmentViewTest.kt
@@ -141,7 +141,7 @@ class AttachmentViewTest {
         }
 
         assertEquals(null, AttachmentCache.bytes("att-2", load))
-        assertEquals(9, AttachmentCache.bytes("att-2", load)?.first())
+        assertEquals(9.toByte(), AttachmentCache.bytes("att-2", load)?.first())
         assertEquals(2, calls)
     }
